% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_replace.R
\name{auto_replace}
\alias{auto_replace}
\title{Apply a replacement function to a \code{rules.frm}.}
\usage{
auto_replace(frm, replacement.f, filter = F)
}
\arguments{
\item{frm}{A \code{data.frame} with columns \code{If}, \code{From}, and \code{To}.}

\item{replacement.f}{A function for transforming the \code{To} column.}

\item{filter}{Logical.  If \code{TRUE} will only apply to rows where \code{From} and \code{To} are different.}
}
\value{
A \code{data.frame} like \code{frm} but with the \code{To} column transformed by \code{replacement.f}.
}
\description{
Several function factories have been implemented to create replacement functions
(\code{hashing_replacement.f}, \code{random_replacement.f}, \code{all_random_replacement.f})
}
\examples{

replace_by <- random_replacement.f()
auto_replace(raw_redaction_rules, replacement.f = replace_by)

}
