% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_error.R
\name{base_error}
\alias{base_error}
\title{A custom abort function}
\usage{
base_error(subclass, message, ..., call)
}
\arguments{
\item{subclass}{a vector of inherited error class}

\item{message}{The error message to display}

\item{...}{Additional arguments to pass to `abort()`}

\item{call}{The call environment to use for the error (defaults to the caller's environment)}
}
\value{
An error object with the specified message and classes
}
\description{
A custom abort function
}
\keyword{internal}
