% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_model_location}
\alias{browse_model_location}
\title{Browse user to folder for UDPipe models.}
\usage{
browse_model_location()
}
\description{
Intended for usage in `udpipe` fails to download automatically.
}
