% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_udpipe_repo}
\alias{browse_udpipe_repo}
\title{Open github link to the 'english-ewt-2.5' UD model.}
\usage{
browse_udpipe_repo(model = "english-ewt")
}
\arguments{
\item{model}{A string naming a UDPipe model.  See `udpipe::udpipe_download_model()` for the list of available models.}
}
\description{
Intended for user download where `udpipe` fails to download automatically.
}
