% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_redact_factory.R
\name{efficient_redact_factory}
\alias{efficient_redact_factory}
\title{Stateful recoding template function}
\usage{
efficient_redact_factory(redact.function)
}
\arguments{
\item{redact.function}{A single input function to perform variable recoding.}
}
\value{
A function
}
\description{
Dynamic programming wrapper to a mono-variable function - performs
given recoding function on unique values and recycles previous states.
}
