% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{List all files in a directory with a given extension}
\usage{
find_files(path, extensions = get_implemented_extensions(), .lis = list())
}
\arguments{
\item{path}{The path to search for files}

\item{extensions}{A character vector of file extensions to search for}

\item{.lis}{A list of file paths}
}
\value{
A list of file paths
}
\description{
List all files in a directory with a given extension
}
\keyword{internal}
