% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_supported_files.R
\name{find_supported_files}
\alias{find_supported_files}
\title{Read all supported files in a folder into a named list of data.frames}
\usage{
find_supported_files(
  data_path,
  extensions = get_implemented_extensions(),
  verbose = FALSE
)
}
\arguments{
\item{data_path}{The file path at which data is stored}

\item{extensions}{Optional.  The set of file extensions to scanned for.}

\item{verbose}{Boolean flag - if TRUE will...}
}
\description{
Read all supported files in a folder into a named list of data.frames
}
\keyword{internal}
