% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_context}
\alias{get_context}
\title{NB: to set the context window size, use \code{set_context_window()}.}
\usage{
get_context(
  sentence,
  token,
  context_window = getOption("pid_pos_context_window")
)
}
\arguments{
\item{sentence}{A character vector of sentences.}

\item{token}{A character vector of tokens.}

\item{context_window}{The width of window around the token to be taken.}
}
\description{
NB: to set the context window size, use \code{set_context_window()}.
}
\keyword{internal}
