% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper_nouns.R
\name{get_proper_nouns}
\alias{get_proper_nouns}
\title{This is a generic method that dispatches based on the first argument.}
\usage{
get_proper_nouns(pos_frm)
}
\arguments{
\item{pos_frm}{The data to be tagged for part of sentence}
}
\description{
This is a generic method that dispatches based on the first argument.
}
