% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_replace.R
\name{hashing_replacement.f}
\alias{hashing_replacement.f}
\title{Function factory for hashing replacement.}
\usage{
hashing_replacement.f(key, salt = "", hash = sha256)
}
\arguments{
\item{key}{The hash key (passed to `hash`)}

\item{salt}{The hash salt}

\item{hash}{The desired hash function (default is `sha256` from `openssl` package).
NB: other functions can be used, if they take `key` as a kew word argument.}
}
\value{
`function`
}
\description{
Function factory for hashing replacement.
}
\examples{
replace_by <- hashing_replacement.f(key="PIDPOS", salt="SALT")
auto_replace(raw_redaction_rules, replacement.f = replace_by)


}
