% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_frame.R
\name{load_replacement_rules.character}
\alias{load_replacement_rules.character}
\title{Convert a file of replacements (as defined with `report_as_rules_template`) to
a function that can be used to modify a data frame.}
\usage{
\method{load_replacement_rules}{character}(object, parse = F)
}
\arguments{
\item{parse}{Binary-flag.  If True the replacement function is parsed.}

\item{path}{File path for read.}
}
\description{
Convert a file of replacements (as defined with `report_as_rules_template`) to
a function that can be used to modify a data frame.
}
