% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longer_than.R
\name{longer_than}
\alias{longer_than}
\title{Produces a function that checks if any strings in a vector is longer than 'limit'}
\usage{
longer_than(limit = 5)
}
\arguments{
\item{limit}{The length to check against}
}
\description{
Produces a function that checks if any strings in a vector is longer than 'limit'
}
\keyword{internal}
