% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_col.R
\name{make_unique_col}
\alias{make_unique_col}
\title{Generate a unique column name for a data frame.}
\usage{
make_unique_col(df, base = ".row_index")
}
\arguments{
\item{df}{A data frame to check for existing column names.}

\item{base}{A character string to use as the base for the column name. Default is ".row_index".}
}
\value{
A character string that is a unique column name for the data frame.
}
\description{
This function generates a unique column name for a data frame by appending a number to a base name until it finds a name that is not already in the data frame.
}
\keyword{internal}
