% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_replacements.R
\name{merge_redactions}
\alias{merge_redactions}
\title{Remove PID from a data frame via a merge/}
\usage{
merge_redactions(frm, cached_redactions, preprocess = utf8_encode)
}
\arguments{
\item{frm}{The data frame to be redacted}

\item{cached_redactions}{A data frame with `If` and `Then` columns}

\item{preprocess}{A function of preprocessing steps to be applied to the text columns.}
}
\description{
Remove PID from a data frame via a merge/
}
