% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_redacter.R
\name{parse_redacter}
\alias{parse_redacter}
\title{Parse a data frame into a redaction function with optional caching.}
\usage{
parse_redacter(redacter, with_cache = T)
}
\arguments{
\item{redacter}{A data.frame containing \code{From}, \code{To} and \code{If} or a file path to}

\item{with_cache}{A binary flag to control if memoization is required.}
}
\description{
Parse a data frame into a redaction function with optional caching.
}
