% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pid_pos.R
\name{pid_pos}
\alias{pid_pos}
\title{Proper Noun Detection}
\usage{
pid_pos(
  frm,
  tagger = "english-ewt",
  filter_func = filter_to_proper_nouns,
  chunk_size = 100,
  to_ignore = c(),
  warn_if_missing = FALSE
)
}
\arguments{
\item{frm}{A data frame to check for proper nouns}

\item{tagger}{Either a string naming a UDPipe model (see  \link[udpipe:udpipe_download_model]{udpipe::udpipe_download_model}
for the list of models) or a custom tagging function (see \href{../articles/custom-functions.html}{Custom Functions}
for details of what is required).}

\item{filter_func}{A function to filter the tagged instances. See
the 'Custom Filtering Functions' section of \code{vignette("custom-functions")}
for more details.}

\item{chunk_size}{The number of sentences to tag at a time.  The optimal value
has yet to be determined.}

\item{to_ignore}{A vector of column names to be ignored by the algorithm.
Intended to be used for variables that are giving strong false positives, such as
IDs or ICD-10 codes.}

\item{warn_if_missing}{Raise a warning if the \code{to_ignore} columns are
not in the data frame.}
}
\value{
A \code{pid_report} (inheriting from tibble) containing:
\itemize{
\item \code{ID}: The location of the sentence in the data frame in the form \verb{Col:<colname> Row:<rownumber>}.
\item \code{Token}: The detected proper noun.
\item \code{Sentence}: The sentence in which the proper noun occurs.
\item \code{Document}: The source string (data frame cell) containing the sentence.
\item \code{Repeats}: The number of times the \code{Document} occurs in the data frame.
\item \verb{Affected Columns}: The columns in the data frame where the \code{Document} occurs.
}
If no proper nouns are detected, an empty data frame is returned.
}
\description{
For a given data set, the function reports each detected instance of a proper
noun and reports the location in the data set, the \code{Document} containing the
proper noun, and how often the \code{Document} occurs.
}
\examples{
data(the_one_in_massapequa)
example.data <- head(the_one_in_massapequa, 50)
try(
  pid_pos(example.data, to_ignore = c("scene", "utterance"))
)

pid_pos(example.data, to_ignore = c("scene", "utterance"), tagger = "english-gum")

tag_ewt <- udpipe_factory("english-ewt")
pid_pos(example.data, to_ignore = c("scene", "utterance"), tagger = tag_ewt)


filter_to_long_proper_nouns <- function(frm) {
  frm |>
    dplyr::filter(nchar(Token) > 1)
  filter_to_proper_nouns(frm)
}

pid_pos(example.data,
  to_ignore = c("scene", "utterance"),
  tagger = tag_ewt, filter = filter_to_long_proper_nouns
)

}
