% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_replacements.R
\name{prepare_redactions}
\alias{prepare_redactions}
\title{Prepare redaction function from replacement rules.}
\usage{
prepare_redactions(object)
}
\arguments{
\item{object}{The `replacement_rules` (can be a path to a csv file
or a `data.frame`).}
}
\description{
Prepare redaction function from replacement rules.
}
