% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proper_noun_report.R
\name{proper_noun_report}
\alias{proper_noun_report}
\title{Detect examples of proper nouns in free text via natural language processing}
\usage{
proper_noun_report(text)
}
\arguments{
\item{text}{A passage of free text or vector of free text.}
}
\value{
list with `Proper Noun`, `All Tags` and `Raw Text` data frames
}
\description{
Given a passage of text or a vector of multiple documents - perform part of speech (POS)
tagging and return any example of proper nouns with context.  Default POS taging is via
the udpipe package and the 'English EWT 2.5' library
}
