% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_redaction_rules}
\alias{raw_redaction_rules}
\title{raw_redaction_rules
An example of a redaction rules produced by the `data_frame_report` function.
It is made using the first 20 rows of `the_one_in_massapequa` data set.}
\format{
An object of class \code{pid_report} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
raw_redaction_rules
}
\description{
\describe{
 \item{If}{The text to be redacted (character)}  
 \item{From}{The text to be replaced (character)}
 \item{To}{The text to replace it with (character)}
}
}
\keyword{datasets}
