% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Reads a data file}
\usage{
read_data(file_path, ...)
}
\arguments{
\item{file_path}{The path to the file to read}

\item{...}{Additional arguments to pass to the read function}
}
\description{
Reads a data file
}
\keyword{internal}
