% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redaction_function_factory.R
\name{redaction_function_factory}
\alias{redaction_function_factory}
\title{Replacement rules to redaction function}
\usage{
redaction_function_factory(rules.frm)
}
\arguments{
\item{rules.frm}{A data.frame with columns `If`, `From` and `To`.}
}
\description{
Convert a `data.frame` of redaction rules into a function that can be applied to a character vector.
}
\examples{
data(the_one_in_massapequa)
example.data <- head(the_one_in_massapequa)

raw_rules <- pid_pos(example.data) |>
  report_to_redaction_rules()

redaction_rules <- auto_replace(raw_rules,
  replacement.f = random_replacement.f()
)

redaction_func <- redaction_function_factory(redaction_rules)

redaction_func(example.data)

}
