% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_reader.R
\name{reinstate_default_reader}
\alias{reinstate_default_reader}
\title{Reinstate the default read functionality for csv, tsv, xls, and xlsx files.}
\usage{
reinstate_default_reader(env = pid.pos_env)
}
\description{
Reinstate the default read functionality for csv, tsv, xls, and xlsx files.
}
\examples{
reinstate_default_reader()
}
