% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_frame.R
\name{report_as_rules_template}
\alias{report_as_rules_template}
\title{Convert the output of `data_frame_report` to a csv file for editing.
It is intended the user changes the `To` column to define desired 
replacements.}
\usage{
report_as_rules_template(report, path = NULL)
}
\arguments{
\item{report}{The 'report' generated by `data_frame_report`}

\item{path}{File path for write.}
}
\description{
Convert the output of `data_frame_report` to a csv file for editing.
It is intended the user changes the `To` column to define desired 
replacements.
}
