% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_on_folder.R
\name{report_on_folder}
\alias{report_on_folder}
\title{Folder Report}
\usage{
report_on_folder(
  data_path,
  report_dir = "Proper Noun Reports",
  tagger = "english-ewt",
  filter_func = filter_to_proper_nouns,
  chunk_size = 100,
  to_ignore = c(),
  export_function = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data_path}{The file path at which data is stored}

\item{report_dir}{The location to write PID reports to}

\item{tagger}{[optional] Either a string naming a UDPipe model (see  \link[udpipe]{udpipe_download_model} 
for the list of models) or a custom tagging function (see \code{vignette("custom-functions")}
for details of what is required).}

\item{filter_func}{[optional] A function to filter the tagged instances. See
the 'Custom Filtering Functions' section of \code{vignette("custom-functions")}
for more details.}

\item{chunk_size}{[optional] The number of sentences to tag at a time.  The optimal value
has yet to be determined.}

\item{to_ignore}{[optional] A vector of column names to be ignored by the algorithm.
Intended to be used for variables that are giving strong false positives, such as
IDs or ICD-10 codes.}

\item{export_function}{A function to control exporting the reports to disk.  Current
options  are `export_as_tree` and `export_flat`}

\item{verbose}{Boolean flag - if TRUE will...}
}
\description{
Iterates over a folder of data files and produces a proper noun report for each.
The reports are saved in the specified `report directory`.
}
\examples{
{
  input_dir <- withr::local_tempdir()
  output_dir <- withr::local_tempdir()

  dir.create(input_dir, recursive = TRUE, showWarnings = FALSE)
  dir.create(output_dir, recursive = TRUE, showWarnings = FALSE)

  example_data <- data.frame(
    text = "Joey went to London",
    stringsAsFactors = FALSE
  )

  utils::write.csv(example_data,
    file.path(input_dir, "example.csv"),
    row.names = FALSE
  )

  paths <- report_on_folder(input_dir, report_dir = output_dir)

  paths
}

}
