% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_frame.R
\name{report_to_replacement_rules}
\alias{report_to_replacement_rules}
\title{Convert the output of `data_frame_report` to a csv file for editing.
It is intended the user changes the `To` column to define desired
replacements.}
\usage{
report_to_replacement_rules(report, path = NULL, add_context = F)
}
\arguments{
\item{report}{The `data.frame` generated by `data_frame_report`}

\item{path}{File path for write.}

\item{add_context}{Binary-flag. If True, a}
}
\value{
data.frame
}
\description{
Convert the output of `data_frame_report` to a csv file for editing.
It is intended the user changes the `To` column to define desired
replacements.
}
