% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentence_frm}
\alias{sentence_frm}
\title{A short data frame of free text including PID.  Used for basic examples
and tests.}
\format{
A data frame with 5 rows and 4 columns:
\describe{
  \item{ID}{The row number (integer)}
  \item{Sentence}{The free text to detect PID in.}
  \item{Numeric}{Example numeric data (discrete) to be ignored by the algorithm}
  \item{Random}{Example numeric data (continuous) to be ignored by the algorithm}
}
}
\usage{
sentence_frm
}
\description{
A short data frame of free text including PID.  Used for basic examples
and tests.
}
\keyword{datasets}
