% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{set_context_window}
\alias{set_context_window}
\title{Set the context window size for the `get_context` function.}
\usage{
set_context_window(x)
}
\arguments{
\item{x}{An integer specifying the number of characters to include
before and after the token in the context.}
}
\description{
Set the context window size for the `get_context` function.
}
\keyword{internal}
