% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_frame.R
\name{template_to_rules}
\alias{template_to_rules}
\title{Convert a 'replacement tempalte' as defined with `report_as_rules_template` to
a function that can be used to modify a data frame.}
\usage{
template_to_rules(path, parse = F)
}
\arguments{
\item{path}{File path for read.}

\item{parse}{Binary-flag.  If True the replacement function is parsed.}
}
\description{
Convert a 'replacement tempalte' as defined with `report_as_rules_template` to
a function that can be used to modify a data frame.
}
