% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_one_in_massapequa}
\alias{the_one_in_massapequa}
\title{The One in Massapequa}
\format{
A data frame with 3,100 rows and 6 columns:
\describe{
  \item{scene}{The scene number (integer)}
  \item{utterance}{The utterance-by-scene number (integer)}
  \item{speaker}{The speaker of the utterance (character)}
  \item{text}{The text of the utterance (character)}
}
}
\usage{
the_one_in_massapequa
}
\description{
A subset of the 'Friends' data set containing the scene, utterance, speaker and
text of the episode 'The One in Massapequa' (s8e18).
}
\keyword{datasets}
