% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_replace.R
\name{auto_replace}
\alias{auto_replace}
\title{Apply a replacement function to a `rules.frm`.}
\usage{
auto_replace(frm, replacement.f, filter = F)
}
\arguments{
\item{frm}{A `data.frame` with columns `If`, `From`, and `To`.}

\item{replacement.f}{A function for transforming the `To` column.}

\item{filter}{Logical.  If `TRUE` will only apply to rows where `From` and `To` are different.}
}
\value{
A `data.frame` like `frm` but with the `To` column transformed by `replacement.f`.
}
\description{
Several function factories have been implemented to create replacement functions
(`hashing_replacement.f`, `random_replacement.f`, `all_random_replacement.f`)
}
\examples{

replace_by <- random_replacement.f()
auto_replace(raw_redaction_rules, replacement.f = replace_by)

}
