% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_report.R
\name{data_frame_report}
\alias{data_frame_report}
\title{Proper Noun Detection}
\usage{
data_frame_report(frm, chunk_size = 100, to_remove = c())
}
\arguments{
\item{frm}{A data frame to check for proper nouns}

\item{chunk_size}{[optional] The number of sentences to tag at a time.  The optimal value
has yet to be determined.}

\item{to_remove}{[optional] A character vector of column names to remove from the data frame}
}
\value{
A tibble containing:
\itemize{
  \item `ID`: The location of the sentence in the data frame in the form `Col:<colname> Row:<rownumber>`.
  \item `Token`: The detected proper noun.
  \item `Sentence`: The sentence containing the proper noun.
  \item Repats: The number of times the sentence occurs in the data frame.
  \item `Affected Columns`: The columns in the data frame where the sentence occurs.
}
If no proper nouns are detected, an empty data frame is returned.
}
\description{
For a given data set, the function reports each detected instance of a proper
noun and reports the location in the data set, the sentence containing the
proper noun, and how often the sentence occurs.
}
\examples{
\dontrun{
example.data <- head(the_one_in_massapequa, 10)
data_frame_report(example.data, to_remove=c("scene", "utterance"))
}
}
