% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_report.R
\name{data_frame_report}
\alias{data_frame_report}
\title{Tags a data frame with part of speech tags.}
\usage{
data_frame_report(frm, chunk_size = 100, to_remove = c())
}
\arguments{
\item{frm}{A data frame to tag}

\item{chunk_size}{The number of sentences to tag at a time}

\item{to_remove}{A character vector of column names to remove from the data frame}
}
\description{
Tags a data frame with part of speech tags.
}
