% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_redact_factory.R
\name{efficient_redaction}
\alias{efficient_redaction}
\title{A wrapper for efficient redaction.}
\usage{
efficient_redaction(frm, redact, n = NULL, .progress = T)
}
\arguments{
\item{frm}{The data frame to be redacted}

\item{redact}{A  function which converts free text to redacted text.}

\item{n}{The number of chunks to split the data frame into for processing.}

\item{.progress}{Whether to show a progress bar.}
}
\description{
A wrapper for efficient redaction.
}
