% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_replacement.R
\name{frame_replacement}
\alias{frame_replacement}
\title{Remove PID from a data frame.}
\usage{
frame_replacement(frm, rules.frm)
}
\arguments{
\item{frm}{The data frame containing text}

\item{rules.frm}{The `data.frame` containing `If`, `From` and `To` rules.}
}
\value{
`data.frame`
}
\description{
Applied the replacement rules (as defined in a `data.frame` with columns
`If`, `From` and `To`) to all character columns in a data frame.
}
