% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_context}
\alias{get_context}
\title{Get the context of a token in a sentence.}
\usage{
get_context(sentences, tokens)
}
\arguments{
\item{sentences}{A character vector of sentences.}

\item{tokens}{A character vector of tokens.}
}
\description{
NB: to set the context window size, use `set_context_window()`.
}
\keyword{internal}
