% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_frame.R
\name{load_replacement_rules}
\alias{load_replacement_rules}
\title{Convert a file of replacements (as defined with `report_as_rules_template`) to
a function that can be used to modify a data frame.}
\usage{
load_replacement_rules(path, parse = F)
}
\arguments{
\item{path}{File path for read.}

\item{parse}{Binary-flag.  If True the replacement function is parsed.}
}
\description{
Convert a file of replacements (as defined with `report_as_rules_template`) to
a function that can be used to modify a data frame.
}
