% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_tagger.R
\name{longer_than}
\alias{longer_than}
\title{Produces a function that checks if any strings in a vector is longer than 'limit'}
\usage{
longer_than(limit = 5)
}
\arguments{
\item{limit}{The length to check against}
}
\description{
Produces a function that checks if any strings in a vector is longer than 'limit'
}
\examples{
longer_than_5 <- longer_than(5)
.str <- c("ABCDE", "ABC", "ABCDEFG")
longer_than_5(.str)

}
