% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact.R
\name{redact}
\alias{redact}
\title{Redact PID}
\usage{
redact(object, redacter, in_batches = T, ...)
}
\arguments{
\item{object}{The object to be redacted - either a vector or data frame}

\item{redacter}{A \code{data.frame} of redaction rules or a function created by \code{redaction_function_factory()}.}

\item{in_batches}{Boolean flag - if True the supplied data will be processed in chunks.}

\item{...}{Other arguments to control batching. See ... for details.}
}
\description{
Redact PID
}
