% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_replacement.R
\name{redaction_function_factory}
\alias{redaction_function_factory}
\title{Create a redaction function from a `data.frame` of replacement rules.}
\usage{
redaction_function_factory(rules.frm)
}
\arguments{
\item{rules.frm}{A data.frame with columns `If`, `From` and `To`.}
}
\description{
Create a redaction function from a `data.frame` of replacement rules.
}
