% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_on_folder.R
\name{report_on_folder}
\alias{report_on_folder}
\title{Itterates over a folder of data files and produces a report on proper nouns.}
\usage{
report_on_folder(
  data_path,
  report_dir = "Proper Noun Reports",
  to_remove = c()
)
}
\arguments{
\item{data_path}{The path to the data files}

\item{report_dir}{The directory to save the reports}

\item{to_remove}{A character vector of column names to remove from the data frame}
}
\description{
Itterates over a folder of data files and produces a report on proper nouns.
}
