% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_frame.R
\name{report_to_redaction_rules}
\alias{report_to_redaction_rules}
\title{Initialize redaction rules}
\usage{
report_to_redaction_rules(report, path = NULL, add_context = F)
}
\arguments{
\item{report}{The `data.frame` generated by `data_frame_report`}

\item{path}{[optional] File path for write.}

\item{add_context}{[optional] Binary-flag. If True, adds a `Context` column giving a subset
of the sentence being redacted around the token.}
}
\value{
A tibble with columns:
\itemize{
\item `If`: The condition (sentence) under which the replacement is to be made.
\item `From`: The original token.
\item `To`: The replacement token.
}
}
\description{
Convert the output of `data_frame_report` to a csv file for editing.
It is intended the user changes the `To` column to define desired
redactions.
}
\examples{
\dontrun{
example.data <- head(the_one_in_massapequa)
report <- data_frame_report(example.data, to_remove = "speaker")
report_to_redaction_rules(report)
}

}
